% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fib}
\alias{fib}
\title{The n-th Fibonacci number}
\usage{
fib(n)
}
\arguments{
\item{n}{The index of the Fibonacci number to retun}
}
\description{
Returns the n-th Fibonacci number.
}
\examples{
fib(4)
}
